+++ src/main/java/com/wellness/retreat/scheduler/controllers/RetreatController.java
@@ -2,10 +2,13 @@ package com.wellness.retreat.scheduler.controllers;
 
 import com.wellness.retreat.scheduler.models.dtos.RetreatDTO;
 import com.wellness.retreat.scheduler.services.RetreatService;
+import jakarta.validation.Valid;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
@@ -20,7 +23,7 @@ public class RetreatController {
     }
 
     @PostMapping
-    public ResponseEntity<RetreatDTO> createRetreat(@RequestBody RetreatDTO retreatDTO) {
+    public ResponseEntity<RetreatDTO> createRetreat(@Valid @RequestBody RetreatDTO retreatDTO) {
         RetreatDTO savedRetreat = retreatService.saveRetreat(retreatDTO);
         return ResponseEntity.ok(savedRetreat);
     }
@@ -36,8 +39,16 @@ public class RetreatController {
         return ResponseEntity.ok(retreatService.getAllRetreats());
     }
 
+    @GetMapping("/search")
+    public ResponseEntity<List<RetreatDTO>> searchRetreats(@RequestParam(required = false) String title,
+                                                           @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime startDate,
+                                                           @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime endDate) {
+        List<RetreatDTO> retreats = retreatService.findRetreatsByCriteria(title, startDate, endDate);
+        return ResponseEntity.ok(retreats);
+    }
+
     @PutMapping("/{id}")
-    public ResponseEntity<RetreatDTO> updateRetreat(@PathVariable Long id, @RequestBody RetreatDTO retreatDTO) {
+    public ResponseEntity<RetreatDTO> updateRetreat(@Valid @PathVariable Long id, @RequestBody RetreatDTO retreatDTO) {
         RetreatDTO updatedRetreat = retreatService.saveRetreat(retreatDTO);
         return ResponseEntity.ok(updatedRetreat);
     }
