+++ src/main/java/com/wellness/retreat/scheduler/models/dtos/GuestDTO.java
@@ -6,11 +6,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.validation.annotation.Validated;
+
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@Validated
 public class GuestDTO {
     private Long id;
     @NotBlank(message = "Invalid first name")
