+++ src/main/java/com/wellness/retreat/scheduler/repositories/RetreatRepository.java
@@ -3,6 +3,7 @@ package com.wellness.retreat.scheduler.repositories;
 import com.wellness.retreat.scheduler.models.entities.Retreat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,5 +13,10 @@ public interface RetreatRepository extends JpaRepository<Retreat, Long> {
 
     @Query
     List<Retreat> findRetreatsBetweenDates(LocalDateTime start, LocalDateTime end);
+
+    List<Retreat> findRetreatsByCriteria(@Param("title") String title,
+                                         @Param("startDate") LocalDateTime startDate,
+                                         @Param("endDate") LocalDateTime endDate);
 }
+
