+++ src/main/java/com/wellness/retreat/scheduler/services/BookingService.java
@@ -1,58 +1,16 @@
 package com.wellness.retreat.scheduler.services;
 
 import com.wellness.retreat.scheduler.models.dtos.BookingDTO;
-import com.wellness.retreat.scheduler.models.entities.Booking;
-import com.wellness.retreat.scheduler.repositories.BookingRepository;
-import org.springframework.beans.BeanUtils;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
-@Service
-public class BookingService {
+public interface BookingService {
+    BookingDTO saveBooking(BookingDTO bookingDTO);
 
-    private final BookingRepository bookingRepository;
+    BookingDTO getBookingById(Long id);
 
-    public BookingService(BookingRepository bookingRepository) {
-        this.bookingRepository = bookingRepository;
-    }
+    List<BookingDTO> getAllBookings();
 
-    public BookingDTO saveBooking(BookingDTO bookingDTO) {
-        Booking booking = convertToEntity(bookingDTO);
-        Booking savedBooking = bookingRepository.save(booking);
-        return convertToDTO(savedBooking);
-    }
-
-    public Optional<BookingDTO> getBookingById(Long id) {
-        return bookingRepository.findById(id)
-                .map(this::convertToDTO);
-    }
-
-    public List<BookingDTO> getAllBookings() {
-        return bookingRepository.findAll().stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-    }
-
-    public void deleteBooking(Long id) {
-        bookingRepository.deleteById(id);
-    }
-
-    private Booking convertToEntity(BookingDTO bookingDTO) {
-        Booking booking = new Booking();
-        BeanUtils.copyProperties(bookingDTO, booking);
-
-        return booking;
-    }
-
-
-
-    private BookingDTO convertToDTO(Booking booking) {
-        BookingDTO bookingDTO = new BookingDTO();
-        BeanUtils.copyProperties(booking, bookingDTO);
-
-        return bookingDTO;
-    }
+    void deleteBooking(Long id);
 }
+
