+++ src/main/java/com/wellness/retreat/scheduler/services/BookingServiceImpl.java
@@ -1,0 +1,58 @@
+package com.wellness.retreat.scheduler.services;
+
+import com.wellness.retreat.scheduler.models.dtos.BookingDTO;
+import com.wellness.retreat.scheduler.models.entities.Booking;
+import com.wellness.retreat.scheduler.repositories.BookingRepository;
+import org.springframework.beans.BeanUtils;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class BookingServiceImpl implements BookingService {
+
+    private final BookingRepository bookingRepository;
+
+    public BookingServiceImpl(BookingRepository bookingRepository) {
+        this.bookingRepository = bookingRepository;
+    }
+
+    @Override
+    public BookingDTO saveBooking(BookingDTO bookingDTO) {
+        Booking booking = convertToEntity(bookingDTO);
+        Booking savedBooking = bookingRepository.save(booking);
+        return convertToDTO(savedBooking);
+    }
+
+    @Override
+    public BookingDTO getBookingById(Long id) {
+        return bookingRepository.findById(id)
+                .map(this::convertToDTO)
+                .orElse(null);
+    }
+
+    @Override
+    public List<BookingDTO> getAllBookings() {
+        return bookingRepository.findAll().stream()
+                .map(this::convertToDTO)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public void deleteBooking(Long id) {
+        bookingRepository.deleteById(id);
+    }
+
+    private Booking convertToEntity(BookingDTO bookingDTO) {
+        Booking booking = new Booking();
+        BeanUtils.copyProperties(bookingDTO, booking);
+        return booking;
+    }
+
+    private BookingDTO convertToDTO(Booking booking) {
+        BookingDTO bookingDTO = new BookingDTO();
+        BeanUtils.copyProperties(booking, bookingDTO);
+        return bookingDTO;
+    }
+}
