+++ src/main/java/com/wellness/retreat/scheduler/services/RetreatService.java
@@ -6,51 +6,14 @@ import com.wellness.retreat.scheduler.repositories.RetreatRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Service
-public class RetreatService {
+public interface RetreatService {
+    RetreatDTO saveRetreat(RetreatDTO retreatDTO);
+    RetreatDTO getRetreatById(Long id);
+    List<RetreatDTO> getAllRetreats();
+    void deleteRetreat(Long id);
 
-    private final RetreatRepository retreatRepository;
-
-    public RetreatService(RetreatRepository retreatRepository) {
-        this.retreatRepository = retreatRepository;
-    }
-
-    public RetreatDTO saveRetreat(RetreatDTO retreatDTO) {
-        Retreat retreat = convertToEntity(retreatDTO);
-        Retreat savedRetreat = retreatRepository.save(retreat);
-        return convertToDTO(savedRetreat);
-    }
-
-    public RetreatDTO getRetreatById(Long id) {
-        return retreatRepository.findById(id)
-                .map(this::convertToDTO)
-                .orElse(null); // or handle as needed
-    }
-
-    public List<RetreatDTO> getAllRetreats() {
-        return retreatRepository.findAll().stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-    }
-
-    public void deleteRetreat(Long id) {
-        retreatRepository.deleteById(id);
-    }
-
-    private Retreat convertToEntity(RetreatDTO retreatDTO) {
-        Retreat retreat = new Retreat();
-        BeanUtils.copyProperties(retreatDTO, retreat);
-
-        return retreat;
-    }
-
-    private RetreatDTO convertToDTO(Retreat retreat) {
-        RetreatDTO retreatDTO = new RetreatDTO();
-        BeanUtils.copyProperties(retreat, retreatDTO);
-
-        return retreatDTO;
-    }
 }
